Remove code repetition from content scripts #18
