+++ src/content-scripts/linkExposure.content.js
@@ -3,115 +3,14 @@
  * @module webScience.linkExposure.content
  */
 
+import { importMatchPatternSet } from "../matching.js";
+import { parseFacebookLinkShim, parseAmpUrl } from "../linkResolution.js";
+
 // async IIFE wrapper to enable await syntax
 (async function () {
 
     let pageManager = null;
 
-    /**
-     * An optimized object for matching against match patterns. A `MatchPatternSet` can provide
-     * a significant performance improvement in comparison to `RegExp`s, in some instances
-     * greater than 100x. A `MatchPatternSet` can also be exported to an object that uses only
-     * built-in types, so it can be persisted or passed to content scripts in extension storage.
-     *
-     * There are several key optimizations in `MatchPatternSet`:
-     *   * URLs are parsed with the `URL` class, which has native implementation.
-     *   * Match patterns are indexed by hostname in a hash map. Lookups are much faster than
-     *     iteratively advancing and backtracking through a complex regular expression, which
-     *     is how domain matching currently occurs with the `Irregexp` regular expression
-     *     engine in Firefox and Chrome.
-     *   * Match patterns with identical scheme, subdomain matching, and host (i.e., that
-     *     differ only in path) are combined.
-     *   * The only remaining use of regular expressions is in path matching, where expressions
-     *     can be (relatively) uncomplicated.
-     *
-     * Future performance improvements could include:
-     *   * Replacing the path matching implementation to eliminate regular expressions entirely.
-     *   * Replacing the match pattern index, such as by implementing a trie.
-     */
-    class MatchPatternSet {
-        /**
-         * Creates a match pattern set from an array of match patterns.
-         * @param {string[]} matchPatterns - The match patterns for the set.
-         */
-        constructor(matchPatterns) {
-            // Defining the special sets of `<all_url>` and wildcard schemes inside the class so
-            // keeping content scripts in sync with this implementation will be easier
-            this.allUrls = false;
-            this.allUrlsSchemeSet = new Set(["http", "https", "ws", "wss", "ftp", "file", "data"]);
-            this.wildcardSchemeSet = new Set(["http", "https", "ws", "wss"]);
-            this.patternsByHost = { };
-        }
-
-        /**
-         * Compares a URL string to the match patterns in the set.
-         * @param {string} url - The URL string to compare.
-         * @returns {boolean} Whether the URL string matches a pattern in the set.
-         */
-        matches(url) {
-            let parsedUrl;
-            try {
-                parsedUrl = new URL(url);
-            } catch {
-                // If the target isn't a true URL, it certainly doesn't match
-                return false;
-            }
-            // Remove the trailing : from the parsed protocol
-            const scheme = parsedUrl.protocol.substring(0, parsedUrl.protocol.length - 1);
-            const host = parsedUrl.hostname;
-            const path = parsedUrl.pathname;
-
-            // Check the special `<all_urls>` match pattern
-            if(this.allUrls && this.allUrlsSchemeSet.has(scheme))
-                return true;
-
-            // Identify candidate match patterns
-            let candidatePatterns = [ ];
-            // Check each component suffix of the hostname for candidate match patterns
-            const hostComponents = parsedUrl.hostname.split(".");
-            let hostSuffix = "";
-            for(let i = hostComponents.length - 1; i >= 0; i--) {
-                hostSuffix = hostComponents[i] + (i < hostComponents.length - 1 ? "." : "") + hostSuffix;
-                const hostSuffixPatterns = this.patternsByHost[hostSuffix];
-                if(hostSuffixPatterns !== undefined)
-                    candidatePatterns = candidatePatterns.concat(hostSuffixPatterns);
-            }
-
-            // Add match patterns with a wildcard host to the set of candidates
-            const hostWildcardPatterns = this.patternsByHost["*"];
-            if(hostWildcardPatterns !== undefined)
-                candidatePatterns = candidatePatterns.concat(hostWildcardPatterns);
-
-            // Check the scheme, then the host, then the path for a match
-            for(const candidatePattern of candidatePatterns) {
-                if((candidatePattern.scheme === scheme) ||
-                    ((candidatePattern.scheme === "*") && this.wildcardSchemeSet.has(scheme))) {
-                    if(candidatePattern.matchSubdomains ||
-                        (candidatePattern.host === "*") ||
-                        (candidatePattern.host === host)) {
-                        if(candidatePattern.wildcardPath ||
-                            candidatePattern.pathRegExp.test(path))
-                            return true;
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        /**
-         * Imports the match pattern set from an opaque object previously generated by `export`.
-         * @param {exportedInternals} - The previously exported internals for the match pattern set.
-         * @example <caption>Example usage of import.</caption>
-         * // const matchPatternSet1 = new MatchPatternSet([ "*://example.com/*" ]);
-         * // const exportedInternals = matchPatternSet.export();
-         * // const matchPatternSet2 = (new MatchPatternSet([])).import(exportedInternals);
-         */
-        import(exportedInternals) {
-            this.allUrls = exportedInternals.allUrls;
-            this.patternsByHost = exportedInternals.patternsByHost;
-        }
-    }
     /**
      * How often (in milliseconds) to check the page for new links.
      * @constant {number}
@@ -207,52 +106,8 @@
         console.debug("Error: linkExposure content script cannot load RegExps from browser.storage.local.");
         return;
     }
-    const linkMatcher = new MatchPatternSet([]);
-    linkMatcher.import(storedLinkMatcher["webScience.linkExposure.linkMatcher"]);
+    const linkMatcher = importMatchPatternSet(storedLinkMatcher["webScience.linkExposure.linkMatcher"]);
     const urlShortenerRegExp = storedUrlShortenerRegExp["webScience.linkExposure.urlShortenerRegExp"];
-    const ampRegExp = storedAmpRegExp["webScience.linkExposure.ampRegExp"];
-
-    /**
-     * A RegExp for matching URLs that have had Facebook's link shim applied.
-     * @constant {RegExp}
-     */
-    const facebookLinkShimRegExp = /^https?:\/\/l.facebook.com\/l\.php\?u=/;
-
-    /**
-     * Parse a URL from Facebook's link shim, if the shim was applied to the URL.
-     * @param {string} url - A URL that may have Facebook's link shim applied.
-     * @returns {string} If Facebook's link shim was applied to the URL, the unshimmed URL. Otherwise, just the URL.
-     */
-    function parseFacebookLinkShim(url) {
-        if(!facebookLinkShimRegExp.test(url))
-            return url;
-
-        // Extract the original URL from the "u" parameter
-        const urlObject = new URL(url);
-        const uParamValue = urlObject.searchParams.get('u');
-        if(uParamValue === null)
-            return url;
-        return uParamValue;
-    }
-
-    /**
-     * Parse the underlying URL from an AMP cache or viewer URL, if the URL is an AMP cache or viewer URL.
-     * @param {string} url - A URL that may be an AMP cache or viewer URL.
-     * @returns {string} If the URL is an AMP cache or viewer URL, the parsed underlying URL. Otherwise, just the URL.
-     */
-    function parseAmpUrl(url) {
-        if(!ampRegExp.test(url))
-            return url;
-        const parsedAmpUrl = ampRegExp.exec(url);
-        // Reconstruct AMP cache URLs
-        if(parsedAmpUrl.groups.ampCacheUrl !== undefined)
-            return "http" +
-                ((parsedAmpUrl.groups.ampCacheIsSecure === "s") ? "s" : "") +
-                "://" +
-                parsedAmpUrl.groups.ampCacheUrl;
-        // Reconstruct AMP viewer URLs, assuming the protocol is HTTPS
-        return "https://" + parsedAmpUrl.groups.ampViewerUrl;
-    }
 
     /**
      * The time when the page last lost the user's attention, or -1 if the page has never had the user's attention.
